Move tdqm placeholder to _progress_bars module
