+++ photutils/utils/_optional_deps.py
@@ -26,25 +26,3 @@ def __getattr__(name):
         return True
 
     raise AttributeError(f'Module {__name__!r} has no attribute {name!r}.')
-
-
-# Define tqdm as a dummy class if it is not available.
-# This is needed to use tqdm as a context manager with multiprocessing.
-try:
-    from tqdm.auto import tqdm
-except ImportError:
-    class tqdm:  # noqa: N801
-        def __init__(self, *args, **kwargs):
-            pass
-
-        def __enter__(self):
-            return self
-
-        def __exit__(self, *exc):
-            pass
-
-        def update(self, *args, **kwargs):
-            pass
-
-        def set_postfix_str(self, *args, **kwargs):
-            pass
