+++ photutils/utils/_progress_bars.py
@@ -49,3 +49,25 @@ def add_progress_bar(iterable=None, desc=None, total=None, text=False):
                         total=total)  # pragma: no cover
 
     return iterable
+
+
+# Define tqdm as a dummy class if it is not available.
+# This is needed to use tqdm as a context manager with multiprocessing.
+try:
+    from tqdm.auto import tqdm
+except ImportError:
+    class tqdm:  # noqa: N801
+        def __init__(self, *args, **kwargs):
+            pass
+
+        def __enter__(self):
+            return self
+
+        def __exit__(self, *exc):
+            pass
+
+        def update(self, *args, **kwargs):
+            pass
+
+        def set_postfix_str(self, *args, **kwargs):
+            pass
